feat(table): ElasticBF: Cache filterblockReader into block cache
In this update, we will be able to control whether FilterBlockReader is managed by the table or the block cache through a parameter. To this end, we have added a unit test called BloomFilterInCache in db_test, which is similar to the BloomFilter, to test the correctness of Reader's program when it is in or out of the cache.

Closes issue#10